Squashed 'Extern/VFLib/' changes from 0d58754..3181389
3181389 Refine angle/distance calculations
1360b7a Fix order of compositors
0732c8f Lift juce::Rectangle to avoid ambiguous symbol
b25abd8 Put blend modes into pixel procs
a0271c0 Fix colour transfer in image fill
0c790d34 Fix Xcode compile

git-subtree-dir: Extern/VFLib
git-subtree-split: 31813893973f0fce9867c98111d2f53048bd5701

+++ modules/vf_unfinished/graphics/vf_DropShadowStyle.h
@@ -42,10 +42,10 @@ struct DropShadowStyle
   bool            active;
   BlendMode::Type mode;
   Colour          colour;
-  double          opacity;    // [0, 1]
-  double          angle;      // in radians
+  float           opacity;    // [0, 1]
+  float           angle;      // in radians
   int             distance;
-  double          spread;     // [0, 1]
+  float           spread;     // [0, 1]
   int             size;       // [0, 250]
   bool            knockout;
